requirements.txt
Django
SemanticScholar
python-dotenv
langchain_community
neomodel
pandas
gunicorn